+++ src/includes/class-alma-settings.php
@@ -20,19 +20,19 @@ use Alma\API\Entities\FeePlan;
 use Alma\API\Entities\Payment;
 use Alma\API\ParamsError;
 use Alma\API\RequestError;
+use Alma\Woocommerce\Exceptions\Alma_Api_Create_Payments_Exception;
 use Alma\Woocommerce\Exceptions\Alma_Api_Share_Of_Checkout_Accept_Exception;
 use Alma\Woocommerce\Exceptions\Alma_Api_Share_Of_Checkout_Deny_Exception;
 use Alma\Woocommerce\Exceptions\Alma_Api_Soc_Last_Update_Dates_Exception;
+use Alma\Woocommerce\Helpers\Alma_Cart_Helper;
 use Alma\Woocommerce\Helpers\Alma_Encryptor_Helper;
 use Alma\Woocommerce\Helpers\Alma_Fee_Plan_Helper;
+use Alma\Woocommerce\Helpers\Alma_Payment_Helper;
 use Alma\Woocommerce\Helpers\Alma_Settings_Helper as Alma_Helper_Settings;
 use Alma\Woocommerce\Exceptions\Alma_Plans_Definition_Exception;
 use Alma\Woocommerce\Helpers\Alma_General_Helper;
-use Alma\Woocommerce\Models\Alma_Cart;
 use Alma\Woocommerce\Helpers\Alma_Constants_Helper;
-use Alma\Woocommerce\Models\Alma_Payment;
 use Alma\Woocommerce\Helpers\Alma_Internationalization_Helper;
-use Alma\Woocommerce\Exceptions\Alma_Api_Create_Payments_Exception;
 use Alma\Woocommerce\Exceptions\Alma_Exception;
 use Alma\Woocommerce\Exceptions\Alma_Api_Fetch_Payments_Exception;
 use Alma\Woocommerce\Exceptions\Alma_Api_Trigger_Payments_Exception;
@@ -60,7 +60,9 @@ use Alma\Woocommerce\Exceptions\Alma_Api_Share_Of_Checkout_Exception;
  * @property bool keys_validity Flag to indicate id the current keys are working
  * @property string selected_fee_plan Admin dashboard fee_plan in edition mode.
  * @property string test_merchant_id Alma TEST merchant ID
+ * @property string test_merchant_name Alma TEST merchant name
  * @property string live_merchant_id Alma LIVE merchant ID
+ * @property string live_merchant_name Alma LIVE merchant name
  * @property string variable_product_price_query_selector Css query selector
  * @property string variable_product_sale_price_query_selector Css query selector for variable discounted products
  * @property string variable_product_check_variations_event JS event for product variation change
@@ -126,12 +128,21 @@ class Alma_Settings {
 	public $fee_plan_helper;
 
 	/**
+	 * The cart helper.
+	 *
+	 * @var Alma_Cart_Helper
+	 */
+	public $cart_helper;
+
+
+	/**
 	 * Constructor.
 	 */
 	public function __construct() {
 		$this->logger           = new Alma_Logger();
 		$this->encryptor_helper = new Alma_Encryptor_Helper();
 		$this->fee_plan_helper  = new Alma_Fee_Plan_Helper();
+		$this->cart_helper      = new Alma_Cart_Helper();
 
 		$this->load_settings();
 	}
@@ -471,64 +482,6 @@ class Alma_Settings {
 	}
 
 	/**
-	 * Call the payment api and create.
-	 *
-	 * @param string  $order_id The order id.
-	 * @param  FeePlan $fee_plan The fee plan.
-	 *
-	 * @return Payment
-	 * @throws Alma_Api_Create_Payments_Exception Create payment exception.
-	 */
-	public function create_payments( $order_id, $fee_plan ) {
-		try {
-			$model_payment = new Alma_Payment();
-			$payment_type  = $this->get_payment_method( $fee_plan );
-
-			$payload = $model_payment->get_payment_payload_from_order( $order_id, $fee_plan, $payment_type );
-
-			return $this->alma_client->payments->create( $payload );
-		} catch ( \Exception $e ) {
-			$this->logger->error( sprintf( 'Api create_payments, order id "%s" , Api message "%s"', $order_id, $e->getMessage() ) );
-			throw new Alma_Api_Create_Payments_Exception( $order_id, $fee_plan );
-		}
-	}
-
-	/**
-	 * Get the payment method fee plan title.
-	 *
-	 * @param FeePlan $fee_plan The fee plan.
-	 * @return string
-	 *
-	 * @throws Alma_Plans_Definition_Exception Exception.
-	 */
-	public function get_payment_method( $fee_plan ) {
-		if ( $fee_plan->isPayNow() ) {
-			return __( 'Selected payment method : Pay Now with Alma', 'alma-gateway-for-woocommerce' );
-		}
-
-		if ( $fee_plan->isPnXOnly() ) {
-			// translators: %d: number of installments.
-			return sprintf( __( 'Selected payment method : %d installments with Alma', 'alma-gateway-for-woocommerce' ), $fee_plan->getInstallmentsCount() );
-		}
-
-		if ( $fee_plan->isPayLaterOnly() ) {
-			$deferred_months = $fee_plan->getDeferredMonths();
-			$deferred_days   = $fee_plan->getDeferredDays();
-
-			if ( $deferred_days ) {
-				// translators: %d: number of deferred days.
-				return sprintf( __( 'Selected payment method : D+%d deferred  with Alma', 'alma-gateway-for-woocommerce' ), $deferred_days );
-			}
-			if ( $deferred_months ) {
-				// translators: %d: number of deferred months.
-				return sprintf( __( 'Selected payment method : M+%d deferred with Alma', 'alma-gateway-for-woocommerce' ), $deferred_months );
-			}
-		}
-
-		throw new Alma_Plans_Definition_Exception();
-	}
-
-	/**
 	 * Saves settings.
 	 *
 	 * @return void
@@ -587,6 +540,30 @@ class Alma_Settings {
 		} catch ( \Exception $e ) {
 			$this->logger->error( sprintf( 'Api fetch_payment, payment id "%s" , Api message "%s"', $payment_id, $e->getMessage() ) );
 			throw new Alma_Api_Fetch_Payments_Exception( $payment_id );
+		}
+	}
+
+
+	/**
+	 * Create the payment.
+	 *
+	 * @param array   $payload The payload.
+	 * @param string  $order_id The order id.
+	 * @param  FeePlan $fee_plan The fee plan.
+	 *
+	 * @return Payment
+	 *
+	 * @throws Alma_Api_Create_Payments_Exception Create payment exception.
+	 */
+	public function create_payment( $payload, $order_id, $fee_plan ) {
+		try {
+			$this->get_alma_client();
+
+			return $this->alma_client->payments->create( $payload );
+
+		} catch ( \Exception $e ) {
+			$this->logger->error( sprintf( 'Api create_payments, order id "%s" , Api message "%s"', $order_id, $e->getMessage() ) );
+			throw new Alma_Api_Create_Payments_Exception( $order_id, $fee_plan );
 		}
 	}
 
@@ -759,6 +736,19 @@ class Alma_Settings {
 	}
 
 	/**
+	 * Get the merchant name in DB.
+	 *
+	 * @return string
+	 */
+	public function get_active_merchant_name() {
+		if ( $this->get_environment() === 'live' ) {
+			return $this->live_merchant_name;
+		}
+
+		return $this->test_merchant_name;
+	}
+
+	/**
 	 * Get the merchant id in api.
 	 *
 	 * @return void
@@ -776,8 +766,9 @@ class Alma_Settings {
 		if ( ! empty( $this->alma_client ) ) {
 
 			try {
-				$merchant                                    = $this->alma_client->merchants->me();
-				$this->{$this->environment . '_merchant_id'} = $merchant->id;
+				$merchant                                      = $this->alma_client->merchants->me();
+				$this->{$this->environment . '_merchant_id'}   = $merchant->id;
+				$this->{$this->environment . '_merchant_name'} = $merchant->name;
 
 			} catch ( \Exception $e ) {
 				$this->__set( 'keys_validity', 'no' );
@@ -954,7 +945,7 @@ class Alma_Settings {
 		}
 
 		return array_filter(
-			$this->get_eligible_plans_keys( ( new Alma_Cart() )->get_total_in_cents() ),
+			$this->get_eligible_plans_keys( $this->cart_helper->get_total_in_cents() ),
 			function ( $key ) use ( $cart_eligibilities ) {
 				if ( is_array( $cart_eligibilities ) ) {
 					return array_key_exists( $key, $cart_eligibilities );
@@ -975,7 +966,7 @@ class Alma_Settings {
 
 			try {
 				$this->get_alma_client();
-				$this->eligibilities = $this->alma_client->payments->eligibility( Alma_Payment::get_eligibility_payload_from_cart() );
+				$this->eligibilities = $this->alma_client->payments->eligibility( Alma_Payment_Helper::get_eligibility_payload_from_cart() );
 			} catch ( \Exception $error ) {
 				$this->logger->error( $error->getMessage(), $error->getTrace() );
 
@@ -1065,7 +1056,7 @@ class Alma_Settings {
 	 * @return array<array>
 	 */
 	public function get_eligible_plans_for_cart() {
-		$amount = ( new Alma_Cart() )->get_total_in_cents();
+		$amount = $this->cart_helper->get_total_in_cents();
 
 		return array_values(
 			array_map(