+++ src/includes/exceptions/class-alma-incorrect-payment-exception.php
@@ -27,15 +27,13 @@ class Alma_Incorrect_Payment_Exception extends Alma_Exception {
 	 * @param string $payment_id The payment id.
 	 * @param string $order_id The order id.
 	 * @param string $payment_state The payment state.
-	 * @param string $first_instalment_state The first instalment.
 	 */
-	public function __construct( $payment_id, $order_id, $payment_state, $first_instalment_state ) {
+	public function __construct( $payment_id, $order_id, $payment_state ) {
 		$message = sprintf(
-			'State payment error/incorrect, Order id "%s", Payment Id "%s", Payment State "%s", First instalement state "%s" ',
+			'State payment error/incorrect, Order id "%s", Payment Id "%s", Payment State "%s" ',
 			$order_id,
 			$payment_id,
-			$payment_state,
-			$first_instalment_state
+			$payment_state
 		);
 
 		parent::__construct( $message );